Single level optimization approach
* Algorithm successful modified to determine phase compositions as well as
binary parameters in one single optimization (instead of the bi-level approach
which calculates the phase stability separately)
* Does not converge very well; ie for most systems the solver returns invalid
solutions.
* Might be due to arbitrary initial values for the phase stability which are not
compatible with the initial binary params- ie initial point violates constraints
* Could possibly be improved by obtaining INITIAL values for phase separation with
phase stability functions from bi-level approach.- Not convinced this will be
effective, due to the complexity of the function to be optimized.